requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
openai>=1.50.0
httpx>=0.27.0
pydantic==2.5.0
pyyaml==6.0.1
certifi>=2024.0.0
aiofiles==23.2.1
